moved Algolia related functions to different folder
